requirements.txt
langchain
langchain-community
langchain-core
langchain-huggingface
sentence-transformers
huggingface-hub
groq
pinecone-client>=3.0.0
redis
python-dotenv